TC_06.003.01 | add rename organization folder
